Enum sirc::options::RegAlloc

pub enum RegAlloc {
    Stack,
    Linear,
    Graph,
    None,
}

Which register allocator to use

Variants

Stack

The naive stack allocator (spills every value on def and reloads on every use)

Linear

A linear-scan allocator, slower than Self::Stack at compile time but generates code that performs significantly better.

Graph

A graph-coloring allocator, much slower than Self::Linear but generates the best code out of the three.

None

Only used for debug reasons, does not perform register allocation